+++ src/modules/gridStateVectorFileIO_mod.f90
@@ -41,22 +41,22 @@ module gridStateVectorFileIO_mod
   logical :: interpToPhysicsGrid  ! for LAM grid, choose to keep physics variables on their original grid
   character(len=3) :: outputFormat ! output format written by 'gio_writeToFile' can only be 'XDF' or 'RSF'
 
-  type netCDFvarID
-    ! derived type to store variable IDs for NEMO increment netCDF files
-    integer :: lon
-    integer :: lat
-    integer :: lev
-    integer :: TM
-    integer :: SALW
-    integer :: UUW
-    integer :: VVW
-    integer :: SSH
-    integer :: seaice
-    integer :: timeCounter
-    integer :: dateBeg
-    integer :: dateEnd
-    integer :: time
-  end type netCDFvarID
+  ! NEMO variables for netCDF files  
+  integer  , parameter :: dimNemovar = 13
+  character(len=20) :: NEMOvarnames(dimNemovar) = (/'nav_lon'     ,     'nav_lat',     'nav_lev', &
+                                                    'time_counter', 'z_inc_dateb', 'z_inc_datef', &
+                                                    'time'        ,      'bckint',      'bckins', &
+                                                    'bckineta'    , 'bckinseaice',      'bckinu', &
+                                                    'bckinv'/)
+  ! '2D' stands for two-dimensional array: (y,x), '3D': (time,y,x), '4D': (time,depth,y,x)
+  ! 'SC': scalar, 'TI': time, 'DE': depth.
+  character(len=2) :: NEMOvartype(dimNemovar)  = (/    '2D'      ,      '2D'    ,      'DE'    , &
+                                                       'TI'      ,      'SC'    ,      'SC'    , &
+                                                       'SC'      ,      '4D'    ,      '4D'    , &
+                                                       '3D'      ,      '3D'    ,      '4D'    , &
+                                                       '4D'/)
+  integer :: NEMOvarid(dimNemovar)
+  integer, parameter  :: referenceDateNEMO = 19500101 ! reference date for netCDF output files
 
   contains
 
@@ -308,7 +308,7 @@ module gridStateVectorFileIO_mod
 
     ! initialize single precision 3D working copy of statevector for reading file
     call gsv_allocate(statevector_file_r4, 1, hco_file, vco_file,                 &
-                      dateStamp_opt = statevector_out%datestamplist(stepIndex),   &
+                      dateStampList_opt = statevector_out%datestamplist(stepIndex:stepIndex),   &
                       mpi_local_opt = .true., mpi_distribution_opt = 'VarsLevs',  &
                       dataKind_opt = 4, allocHeightSfc_opt = readHeightSfc,       &
                       varNames_opt = varNamesToRead,                              &
@@ -322,7 +322,7 @@ module gridStateVectorFileIO_mod
 
     ! initialize single precision 3D working copy of statevector for horizontal interpolation result
     call gsv_allocate(statevector_hinterp_r4, 1, statevector_out%hco, vco_file,   &
-                      dateStamp_opt = statevector_out%datestamplist(stepIndex),   &
+                      dateStampList_opt = statevector_out%datestamplist(stepIndex:stepIndex),   &
                       mpi_local_opt = .true., mpi_distribution_opt='VarsLevs',    &
                       dataKind_opt = 4, allocHeightSfc_opt = readHeightSfc,       &
                       varNames_opt = varNamesToRead,                              &
@@ -343,7 +343,7 @@ module gridStateVectorFileIO_mod
 
     ! initialize double precision 3D working copy of statevector for mpi communication result
     call gsv_allocate(statevector_tiles, 1, statevector_out%hco, vco_file,     &
-                      dateStamp_opt = statevector_out%datestamplist(stepIndex),&
+                      dateStampList_opt = statevector_out%datestamplist(stepIndex:stepIndex),&
                       mpi_local_opt = .true., mpi_distribution_opt = 'Tiles',  &
                       dataKind_opt = 8, allocHeightSfc_opt = readHeightSfc,    &
                       varNames_opt = varNamesToRead)
@@ -356,7 +356,7 @@ module gridStateVectorFileIO_mod
 
     ! initialize double precision 3D working copy of statevector for mpi communication result
     call gsv_allocate(statevector_vinterp, 1, statevector_out%hco, statevector_out%vco, &
-                      dateStamp_opt = statevector_out%datestamplist(stepIndex), &
+                      dateStampList_opt = statevector_out%datestamplist(stepIndex:stepIndex), &
                       mpi_local_opt = .true., mpi_distribution_opt = 'Tiles', dataKind_opt = 8, &
                       allocHeightSfc_opt = readHeightSfc, varNames_opt = varNamesToRead)
 
@@ -419,7 +419,7 @@ module gridStateVectorFileIO_mod
 
     ! initialize single precision 3D working copy of statevector for reading file
     call gsv_allocate(statevector_file_r4, 1, statevector_out%hco, statevector_out%vco, &
-                      dateStamp_opt = statevector_out%datestamplist(stepIndex),  &
+                      dateStampList_opt = statevector_out%datestamplist(stepIndex:stepIndex),  &
                       mpi_local_opt = .true., mpi_distribution_opt = 'VarsLevs', &
                       dataKind_opt = 4, allocHeightSfc_opt = readHeightSfc,      &
                       varNames_opt = varNamesToRead)
@@ -434,7 +434,7 @@ module gridStateVectorFileIO_mod
 
     !-- 3.0 MPI communication from vars/levels to lat/lon tiles
     call gsv_allocate(statevector_tiles, 1, statevector_out%hco, statevector_out%vco, &
-                      dateStamp_opt = statevector_out%datestamplist(stepIndex), &
+                      dateStampList_opt = statevector_out%datestamplist(stepIndex:stepIndex), &
                       mpi_local_opt = .true., mpi_distribution_opt = 'Tiles',   &
                       dataKind_opt = 8, allocHeightSfc_opt = readHeightSfc,     &
                       varNames_opt = varNamesToRead)
@@ -495,7 +495,7 @@ module gridStateVectorFileIO_mod
 
     ! initialize single precision 3D working copy of statevector for reading file
     call gsv_allocate(statevector_file_r4, 1, hco_file, vco_file,                       &
-                      dateStamp_opt = statevector_out_r4%datestamplist(stepIndex),      &
+                      dateStampList_opt = statevector_out_r4%datestamplist(stepIndex:stepIndex), &
                       mpi_local_opt = .false., dataKind_opt = 4,                        &
                       allocHeightSfc_opt = readHeightSfc, varNames_opt = varNamesToRead,&
                       hInterpolateDegree_opt=statevector_out_r4%hInterpolateDegree,     &
@@ -507,7 +507,7 @@ module gridStateVectorFileIO_mod
 
     ! initialize single precision 3D working copy of statevector for horizontal interpolation result
     call gsv_allocate(statevector_hinterp_r4, 1, statevector_out_r4%hco, vco_file,      &
-                      dateStamp_opt = statevector_out_r4%datestamplist(stepIndex),      &
+                      dateStampList_opt = statevector_out_r4%datestamplist(stepIndex:stepIndex), &
                       mpi_local_opt = .false., dataKind_opt=4,                          &
                       allocHeightSfc_opt = readHeightSfc, varNames_opt = varNamesToRead,&
                       hInterpolateDegree_opt = statevector_out_r4%hInterpolateDegree,   &
@@ -527,7 +527,7 @@ module gridStateVectorFileIO_mod
 
     ! initialize double precision 3D working copy of statevector for mpi communication result
     call gsv_allocate(statevector_vinterp_r4, 1, statevector_out_r4%hco, statevector_out_r4%vco, &
-                      dateStamp_opt=statevector_out_r4%datestamplist(stepIndex),                 &
+                      dateStampList_opt=statevector_out_r4%datestamplist(stepIndex:stepIndex),   &
                       mpi_local_opt=.false., dataKind_opt=4,                                     &
                       allocHeightSfc_opt=readHeightSfc, varNames_opt=varNamesToRead)
 
@@ -652,11 +652,11 @@ module gridStateVectorFileIO_mod
     integer :: imode, ierr, newdate, prntdate, prnttime
     integer :: numberRecords, timeCounterID, dimTimeCounterID
     character(len = nf90_max_name) :: recordDimName
-    integer, parameter  :: referenceDateORCA025 = 19500101 ! reference date for netCDF output files
     integer :: refDateStamp, currentDateStamp
     logical :: foundRequiredState
     integer :: timeIndexToRead
     integer :: nDims, nVars, nGlobalAtts, unlimDimID
+    integer :: numDimsInFile ! returned number of variable dimensions
     
     write(*,*) 'gio_readFileNetCDF: Start reading: ', trim(fileName)
 
@@ -710,8 +710,8 @@ module gridStateVectorFileIO_mod
 				            start = (/            1/), &
                                             count = (/numberRecords/)))
     imode = 3
-    ierr = newdate(refDateStamp, referenceDateORCA025, 0, imode)
-    write(*,*) 'gio_readFileNetCDF: reference date/datestamp: ', referenceDateORCA025, refDateStamp
+    ierr = newdate(refDateStamp, referenceDateNEMO, 0, imode)
+    write(*,*) 'gio_readFileNetCDF: reference date/datestamp: ', referenceDateNEMO, refDateStamp
 
     foundRequiredState = .false.
     do timeIndex = 1, numberRecords    
@@ -753,12 +753,23 @@ module gridStateVectorFileIO_mod
       write(*,*) 'gio_readFileNetCDF: reading varName MIDAS, varName NEMO, levIndex, numLev, stepIndex: ', &
 		 trim(varName), ', ', trim(varNameNetCDF), ', ', levIndex, numLevVar, timeIndexToRead
 
-      call utl_checkNetCDFstatus(nf90_inq_varid(ncid, trim(varNameNetCDF), varID))
+      ! get variable ID from varNameNetCDF
+      call utl_checkNetCDFstatus(nf90_inq_varid(ncid, trim(varNameNetCDF), varid))
+      ! get returned number of dimensions from varid
+      call utl_checkNetCDFstatus(nf90_inquire_variable(ncid, varid, ndims = numDimsInFile))
 
       ! Read a 2D field from file
-      call utl_checkNetCDFstatus(nf90_get_var(ncid, varID, fileField2D, &
-				              start = (/ 1,  1, levIndex, timeIndexToRead/),&
-				              count = (/ni, nj,        1,               1/)))
+      if (numDimsInFile == 3) then
+        ! surface NEMO variable (x, y, time_counter)
+        call utl_checkNetCDFstatus(nf90_get_var(ncid, varID, fileField2D, &
+				                start = (/ 1,  1, timeIndexToRead/),&
+				                count = (/ni, nj,               1/)))
+      else if (numDimsInFile == 4) then
+        ! 3D NEMO variable (x, y, depth, time_counter)
+        call utl_checkNetCDFstatus(nf90_get_var(ncid, varID, fileField2D, &
+	   			                start = (/ 1,  1, levIndex, timeIndexToRead/),&
+				                count = (/ni, nj,        1,               1/)))
+      end if
 
       field_r4_ptr(:,:, varLevIndex, 1) = fileField2D(:,:,1,1)
 
@@ -1455,7 +1466,7 @@ module gridStateVectorFileIO_mod
         ! allocate stateVector for storing just 1 time step
         if (batchIndex == 1) then
           call gsv_allocate(stateVector_1step_r4, 1, stateVectorTrial_ptr%hco,                &
-                            stateVectorTrial_ptr%vco, dateStamp_opt = dateStamp,              &
+                            stateVectorTrial_ptr%vco, dateStampList_opt = (/dateStamp/),          &
                             mpi_local_opt = .false., dataKind_opt = 4,                        &
                             allocHeightSfc_opt=allocHeightSfc, varNames_opt = varNamesToRead, &
                             hInterpolateDegree_opt = stateVectorTrial_ptr%hInterpolateDegree, &
@@ -2583,44 +2594,45 @@ module gridStateVectorFileIO_mod
   !--------------------------------------------------------------------------
   ! gio_writeToFileNetCDF
   !--------------------------------------------------------------------------
-  subroutine gio_writeToFileNetCDF(stateVector_in, fileNameTemplate,     &
+  subroutine gio_writeToFileNetCDF(stateVector_in, fileNameTemplate, validDateStamp, &
                                    stepIndex_opt, containsFullField_opt, &
-                                   varLevIndexBeg_opt, varLevIndexEnd_opt)
+                                   varLevIndexBeg_opt, varLevIndexEnd_opt, &
+                                   timeCounter_opt)
     !
     ! :Purpose: Write a statevector object to a netCDF file.
     !
     implicit none
 
     ! Arguments:
-    type(struct_gsv), target,   intent(in) :: stateVector_in   ! input stateVector object
-    character(len=*),           intent(in) :: fileNameTemplate ! template for increment file name
-    integer,          optional, intent(in) :: stepIndex_opt    ! step index
+    type(struct_gsv), target,   intent(in) :: stateVector_in        ! input stateVector object
+    character(len=*),           intent(in) :: fileNameTemplate      ! template for increment file name
+    integer         ,           intent(in) :: validDateStamp        ! valid dateStamp obtained from stateVector%anltime  
+    integer,          optional, intent(in) :: stepIndex_opt         ! step index
     logical,          optional, intent(in) :: containsFullField_opt ! contains or not full field
-    integer,          optional, intent(in) :: varLevIndexBeg_opt ! start index if writing only a limited set of 'varLev's
-    integer,          optional, intent(in) :: varLevIndexEnd_opt ! end index if writing only a limited set of 'varLev's
-
+    integer,          optional, intent(in) :: varLevIndexBeg_opt    ! start index if writing only a limited set of 'varLev's
+    integer,          optional, intent(in) :: varLevIndexEnd_opt    ! end index if writing only a limited set of 'varLev's
+    integer         , optional, intent(in) :: timeCounter_opt       ! time counter for array in netcdf file (default is stepIndex) 
     ! Locals:
     logical :: iDoWriting, containsFullField
     integer :: ierr, ncid, stepIndex
-    integer :: ni, nj, nk
-    integer :: varID
-    integer :: levIndex, nlev, varIndex, varLevIndex
+    integer :: ni, nj
+    integer :: levIndex, numLev, varIndex, varLevIndex
     integer :: yourid, nsize, youridy, youridx
     integer :: varLevIndexBeg, varLevIndexEnd
     character(len=4), pointer :: varNamesToRead(:)
-    character(len=4) :: nomvar
+    character(len=4) :: varName
     real(4), allocatable :: work2d_r4(:,:)
     real(4), allocatable :: gd_send_r4(:,:), gd_recv_r4(:,:,:)
     real(8), pointer :: field_r8(:,:,:,:)
     real(4), pointer :: field_r4(:,:,:,:)
     type(struct_gsv), pointer :: stateVector
     type(struct_gsv), target  :: stateVector_tiles
-    type(netCDFvarID) :: varIDs
-    integer :: datestampNetCDFend, numberHoursEnd
-    real(8) :: netCDFtimeEnd, netCDFtimeBeg
-    integer(8) :: numberSecondsBeg, numberSecondsEnd
+    integer :: varIndexNEMO
+    real(8) :: netCDFtime
+    integer(8) :: numberSeconds
     character(len=100) :: fileName
-    integer, parameter :: referenceDate = 19500101 ! reference date of NEMO
+    logical :: fileExists
+    integer :: currentDateStamp, timeLevel
     
     call msg('gio_writeToFileNetCDF', 'START')
     
@@ -2662,19 +2674,22 @@ module gridStateVectorFileIO_mod
     end if
     call msg('gio_writeToFileNetCDF', 'containsFullField = '//str(containsFullField))
 
-    if ( present(varLevIndexBeg_opt) .and. present(varLevIndexEnd_opt) ) then
+    if (present(varLevIndexBeg_opt) .and. present(varLevIndexEnd_opt)) then
       varLevIndexBeg = varLevIndexBeg_opt
       varLevIndexEnd = varLevIndexEnd_opt
     else
-      if ( present(varLevIndexBeg_opt) ) then
-        call utl_abort('gio_writeToFile: An argument ''varLevIndexEnd_opt'' must be given when ''varLevIndexBeg_opt_opt'' is given.')
+      if (present(varLevIndexBeg_opt)) then
+        call utl_abort('gio_writeToFileNetCDF: An argument ''varLevIndexEnd_opt'' must be given when ''varLevIndexBeg_opt_opt'' is given.')
       end if
-      if ( present(varLevIndexEnd_opt) ) then
-        call utl_abort('gio_writeToFile: An argument ''varLevIndexBeg_opt'' must be given when ''varLevIndexEnd_opt_opt'' is given.')
+      if (present(varLevIndexEnd_opt)) then
+        call utl_abort('gio_writeToFileNetCDF: An argument ''varLevIndexBeg_opt'' must be given when ''varLevIndexEnd_opt_opt'' is given.')
       end if
-
       varLevIndexBeg = 1
       varLevIndexEnd = gsv_getNumVarLev(stateVector)
+    end if
+
+    if (varLevIndexBeg /= 1 .or. varLevIndexEnd /= gsv_getNumVarLev(stateVector)) then
+      call utl_abort('gio_writeToFileNetCDF: Parallel access to netCDF files is not implemented')
     end if
 
     ! if step index not specified, choose anltime (usually center of window)
@@ -2683,22 +2698,24 @@ module gridStateVectorFileIO_mod
     else
       stepIndex = stateVector%anltime
     end if
-    call msg('gio_writeToFileNetCDF', 'Step index: '//str(stepIndex))
+    
+    ! define time level for netCDF file where to write fields
+    if (present(timeCounter_opt)) then
+      timeLevel = timeCounter_opt
+    else
+      timeLevel = stepIndex
+    end if
 
     ni     = stateVector%ni
     nj     = stateVector%nj
-    nk     = 1
     
     ! get number of levels for NEMO grid, 
     ! which is supposed to be equal for all 3D variables
-    do varIndex = 1, vnl_numvarmax 
-      if (gsv_varExist(statevector, vnl_varNameList(varIndex))) then
-        if (trim(vnl_varNameList(varIndex)) == 'TM') then
-          nlev = statevector%varNumLev(varIndex)
-	  exit
-        end if
-      end if
-    end do
+    if (gsv_varExist(statevector, 'TM')) then
+      numLev = gsv_getNumLevFromVarName(stateVector, 'TM')
+    else
+      call utl_abort('gio_writeToFileNetCDF: variable TM should be present in the state vector.')
+    end if
     
     ! only proc 0 does writing or each proc when data is global 
     ! (assuming only called for proc with global data)
@@ -2709,36 +2726,48 @@ module gridStateVectorFileIO_mod
     !
     if (iDoWriting) then
 
-      ! compute begin and and time for the increments 
-      datestampNetCDFend = stateVector%dateStampList(stepIndex)
-      call msg('gio_writeToFileNetCDF', 'Datestamp END: ' // str(datestampNetCDFend))
+      ! compute begin and and end time for the increments 
+      currentDateStamp = stateVector%dateStampList(stepIndex)
+      call msg('gio_writeToFileNetCDF', 'Current datestamp: ' // str(currentDateStamp))
 
-      ! compute difference in hours between current date (END) and reference
-      call tim_getHoursSinceReferenceDate(datestampNetCDFend, referenceDate, numberHoursEnd)
-      numberSecondsEnd = numberHoursEnd * 3600
-      netCDFtimeEnd = abs(real(numberSecondsEnd, 8))
-      call msg('gio_writeToFileNetCDF', 'Number of hours/netCDF date END  : ' // &
-                                        str(numberHoursEnd) // ', ' // &
-					str(netCDFtimeEnd))
-      ! extract dstepobsinc hours to get hours since reference date for BEGIN, and convert to seconds
-      numberSecondsBeg = (numberHoursEnd - int(tim_dstepobsinc)) * 3600
-      netCDFtimeBeg = abs(real(numberSecondsBeg, 8))
-      call msg('gio_writeToFileNetCDF', 'Number of hours/netCDF date BEGIN: '//&
-                                        str(numberHoursEnd - tim_dstepobsinc)//', '//&
-					str(netCDFtimeBeg))
-
-      !- create increment file, close and reopen it here for writing the increments
       fileName = trim(fileNameTemplate)//'.nc'	
+
+      ! check if the file already exists
+      inquire(file = trim(fileName), exist = fileExists)
+      if (.not. fileExists) then
       
-      call gio_createOceanIncrementNetCDFfile(trim(fileName), ni, nj, nlev, varIDs)
+        !- create increment file and close it
+        call gio_createOceanIncrementNetCDFfile(trim(fileName), ni, nj, numLev)
+
+        ! compute difference in seconds between valid dateStamp date and NEMO reference
+        call tim_getSecondsSinceReferenceDate(validDateStamp, &
+                                              referenceDateNEMO, numberSeconds)
+        netCDFtime = abs(real(numberSeconds, 8))
+	! reopen the file for writing analysis increments and dates
+        call utl_checkNetCDFstatus(nf90_open(trim(fileName), nf90_write, ncid))
+        ! put 'z_inc_dateb', 'z_inc_datef' and 'time' into netCDF file
+        do varIndexNEMO = 5, 7
+          call utl_checkNetCDFstatus(nf90_inq_varid(ncid, NEMOvarnames(varIndexNEMO), NEMOvarid(varIndexNEMO)))
+          call utl_checkNetCDFstatus(nf90_put_var(ncid, NEMOvarid(varIndexNEMO), netCDFtime))
+        end do
+	call utl_checkNetCDFstatus(nf90_close(ncid))
+	
+      else
+        call msg('gio_writeToFileNetCDF', &
+	         'Output file '//trim(fileName)//' has already been created for the first time step')
+      end if
+      
       ! reopen the file for writing analysis increments and dates
       call utl_checkNetCDFstatus(nf90_open(trim(fileName), nf90_write, ncid))
-      ! put time END into netCDF file: time and z_inc_datef variables
-      call utl_checkNetCDFstatus(nf90_put_var(ncid, varIDs%dateEnd, netCDFtimeEnd))
-      call utl_checkNetCDFstatus(nf90_put_var(ncid, varIDs%time   , netCDFtimeEnd))
-      call utl_checkNetCDFstatus(nf90_put_var(ncid, varIDs%dateBeg, netCDFtimeBeg))
+      ! compute difference in seconds between current date and NEMO reference
+      call tim_getSecondsSinceReferenceDate(currentDateStamp, referenceDateNEMO, numberSeconds)
+      netCDFtime = abs(real(numberSeconds, 8))
+      ! put netCDFtime into 'time_counter', varIndexNEMO = 4
+      call utl_checkNetCDFstatus(nf90_inq_varid(ncid, NEMOvarnames(4), NEMOvarid(4)))
+      call utl_checkNetCDFstatus(nf90_put_var(ncid, NEMOvarid(4), netCDFtime, &
+                                              start = (/timeLevel/)))
 
-    end if
+    end if ! iDoWriting
 
     allocate(gd_send_r4(stateVector%lonPerPEmax, stateVector%latPerPEmax))
     if (mmpi_myid == 0 .or. (.not. stateVector%mpi_local)) then
@@ -2756,37 +2785,37 @@ module gridStateVectorFileIO_mod
     end if
 
     do varLevIndex = varLevIndexBeg, varLevIndexEnd
-      nomvar = trim(gsv_getVarNameFromVarLev(stateVector,varLevIndex))
- 
-      call msg('gio_writeToFileNetCDF', 'Writing variable: '//nomvar)
-      select case(trim(nomvar))
-      case('TM')
-        varID = varIDs%TM
-      case('SALW')
-        varID = varIDs%SALW
-      case('UUW')
-        varID = varIDs%UUW
-      case('VVW')
-        varID = varIDs%VVW
-      case('SSH')
-        varID = varIDs%SSH
-      case default
-        call utl_abort('gio_writeToFileNetCDF: requested vnl_varNameList does not exist '//&
-                       trim(nomvar))
-      end select
-      write(*,*) 'gio_writeToFileNetCDF: varID: ', varID , nomvar
 
-      nlev = gsv_getNumLevFromVarName(stateVector,nomvar)
-      levIndex = gsv_getLevFromVarLev(statevector,varLevIndex)
+      varName = trim(gsv_getVarNameFromVarLev(stateVector, varLevIndex))
+
+      select case(trim(varName))
+	case('TM')
+	  varIndexNEMO = 8
+	case('SALW')
+	  varIndexNEMO = 9
+        case('UUW')
+	  varIndexNEMO = 12
+        case('VVW')
+	  varIndexNEMO = 13
+        case('SSH')
+	  varIndexNEMO = 10
+	case default
+	  call utl_abort('gio_writeToFileNetCDF: requested variable name does not exist '//trim(varName))
+      end select
+
+      levIndex = gsv_getLevFromVarLev(stateVector, varLevIndex)
+
+      call msg('gio_writeToFileNetCDF', 'Writing: '//trim(varName)//', level: '//&
+	       str(levIndex)//', timestep: '//str(timeLevel)//', dateStamp: '//str(currentDateStamp)) 
 
       if (stateVector%dataKind == 8) then
-        call gsv_getField(stateVector, field_r8, nomvar)
-        gd_send_r4(1:stateVector%lonPerPE,  &
+        call gsv_getField(stateVector, field_r8, varName)
+        gd_send_r4(1:stateVector%lonPerPE, &
                    1:stateVector%latPerPE) =  &
             real(field_r8(stateVector%myLonBeg : stateVector%myLonEnd, &
                           stateVector%myLatBeg : stateVector%myLatEnd, levIndex, stepIndex), 4)
       else
-        call gsv_getField(stateVector, field_r4, nomvar)
+        call gsv_getField(stateVector, field_r4, varName)
         gd_send_r4(1:stateVector%lonPerPE,  &
                    1:stateVector%latPerPE) =  &
             field_r4(stateVector%myLonBeg : stateVector%myLonEnd, &
@@ -2809,8 +2838,8 @@ module gridStateVectorFileIO_mod
             yourid = youridx + youridy * mmpi_npex
             work2d_r4(stateVector%allLonBeg(youridx + 1) : stateVector%allLonEnd(youridx + 1), &
                       stateVector%allLatBeg(youridy + 1) : stateVector%allLatEnd(youridy + 1)) = &
-                gd_recv_r4(1 : stateVector%allLonPerPE(youridx + 1),  &
-                           1 : stateVector%allLatPerPE(youridy + 1), yourid + 1)
+            gd_recv_r4(1 : stateVector%allLonPerPE(youridx + 1),  &
+                       1 : stateVector%allLatPerPE(youridy + 1), yourid + 1)
           end do
         end do
         !$OMP END PARALLEL DO
@@ -2822,18 +2851,31 @@ module gridStateVectorFileIO_mod
       if (iDoWriting) then
 
         !- Convert Kelvin to Celcius only if full field
-        if (containsFullField .and. trim(nomvar) == 'TT' &
-	                      .or.  trim(nomvar) == 'TM') then
+        if (containsFullField .and. trim(varName) == 'TT' &
+                              .or.  trim(varName) == 'TM') then
           where (work2d_r4(:,:) > 100.0)
             work2d_r4(:,:) = work2d_r4(:,:) - mpc_k_c_degree_offset_r4
           end where
         end if
 
-        !- Writing to file
-        call utl_checkNetCDFstatus(nf90_put_var(ncid, varID, work2d_r4, &
-                                   start = (/ 1,  1, levIndex, 1/),     &
-                                   count = (/ni, nj,        1, 1/)),    &
-                                   'gio_writeToFileNetCDF', 'nf90_put_var')
+        ! inquire current varid
+        call utl_checkNetCDFstatus(nf90_inq_varid(ncid, NEMOvarnames(varIndexNEMO), NEMOvarid(varIndexNEMO)))
+	
+	!- Writing to file
+	if (NEMOvartype(varIndexNEMO) == '3D') then	 
+          call utl_checkNetCDFstatus(nf90_put_var(ncid, NEMOvarid(varIndexNEMO), work2d_r4, &
+                                                  start = (/ 1,  1,  timeLevel/), &
+                                                  count = (/ni, nj,	     1/)), &
+                                     'gio_writeToFileNetCDF', 'nf90_put_var')
+	else if (NEMOvartype(varIndexNEMO) == '4D') then      
+          call utl_checkNetCDFstatus(nf90_put_var(ncid, NEMOvarid(varIndexNEMO), work2d_r4, &
+                                                  start = (/ 1,  1, levIndex, timeLevel/), &
+                                                  count = (/ni, nj,	   1,	      1/)), &
+                                     'gio_writeToFileNetCDF', 'nf90_put_var')
+	else
+	  call utl_abort('gio_writeToFileNetCDF: wrong NEMO vartype for variable: '//&
+                         vnl_varNameList(varIndex)//' ('//trim(NEMOvarnames(varIndexNEMO))//')')
+        end if
 
       end if ! iDoWriting
     end do ! varLevIndex
@@ -2863,7 +2905,7 @@ module gridStateVectorFileIO_mod
   !--------------------------------------------------------------------------
   ! gio_createOceanIncrementNetCDFfile
   !--------------------------------------------------------------------------
-  subroutine gio_createOceanIncrementNetCDFfile(fileName, ni, nj, nk, varIDs)
+  subroutine gio_createOceanIncrementNetCDFfile(fileName, ni, nj, nk)
     !
     ! :Purpose: create an empty netCDF file for ocean-seaice increments
     !           NEMO-CICE requires the following increments:
@@ -2878,15 +2920,14 @@ module gridStateVectorFileIO_mod
 
     ! Arguments:
     character(len=*) , intent(in)  :: fileName   ! file name being created
-    integer          , intent(in)  :: ni, nj, nk ! ORCA025 grid dimensions
-    type(netCDFvarID), intent(out) :: varIDs     ! variable IDs in netCDF file   
+    integer          , intent(in)  :: ni, nj, nk ! NEMO grid dimensions
 
     ! Locals:
-    integer :: ncid
+    integer :: ncid, variableIndex
     integer :: latDimID, lonDimID, levDimID, timeID
-    integer :: dimID2D(2), dimID3D(3), dimID4D(4) 
+    integer :: dimID2D(2), dimID3D(3), dimID4D(4), dimIDTI(1), dimIDDE(1)
 
-    call msg('gio_createOceanIncrementNetCDFfile', 'Creating file: '//str(trim(fileName)))
+    call msg('gio_createOceanIncrementNetCDFfile', 'Creating file: '//trim(fileName))
     ! create netCDF file    
     call utl_checkNetCDFstatus(nf90_create(trim(fileName), nf90_netcdf4, ncid))
 
@@ -2899,21 +2940,31 @@ module gridStateVectorFileIO_mod
     dimID4D = (/lonDimID, latDimID, levDimID, timeID/)	
     dimID3D = (/lonDimID, latDimID,           timeID/)
     dimID2D = (/lonDimID, latDimID                  /)
-      	
-    call utl_checkNetCDFstatus(nf90_def_var(ncid, 'nav_lon'     , nf90_float , dimID2D , varIDs%lon))
-    call utl_checkNetCDFstatus(nf90_def_var(ncid, 'nav_lat'     , nf90_float , dimID2D , varIDs%lat))
-    call utl_checkNetCDFstatus(nf90_def_var(ncid, 'nav_lev'     , nf90_float , levDimID, varIDs%lev))
-    call utl_checkNetCDFstatus(nf90_def_var(ncid, 'time_counter', nf90_double, timeID  , varIDs%timeCounter))
-    call utl_checkNetCDFstatus(nf90_def_var(ncid, 'z_inc_dateb' , nf90_double,           varIDs%dateBeg))
-    call utl_checkNetCDFstatus(nf90_def_var(ncid, 'z_inc_datef' , nf90_double,           varIDs%dateEnd))
-    call utl_checkNetCDFstatus(nf90_def_var(ncid, 'time'        , nf90_double,           varIDs%time))
-
-    call utl_checkNetCDFstatus(nf90_def_var(ncid, 'bckint'     , nf90_double, dimID4D , varIDs%TM))
-    call utl_checkNetCDFstatus(nf90_def_var(ncid, 'bckins'     , nf90_double, dimID4D , varIDs%SALW))
-    call utl_checkNetCDFstatus(nf90_def_var(ncid, 'bckineta'   , nf90_double, dimID3D , varIDs%SSH))
-    call utl_checkNetCDFstatus(nf90_def_var(ncid, 'bckinseaice', nf90_double, dimID3D , varIDs%seaice))
-    call utl_checkNetCDFstatus(nf90_def_var(ncid, 'bckinu'     , nf90_double, dimID4D , varIDs%UUW))
-    call utl_checkNetCDFstatus(nf90_def_var(ncid, 'bckinv'     , nf90_double, dimID4D , varIDs%VVW))
+    dimIDTI = (/                              timeID/)
+    dimIDDE = (/                    levDimID        /)
+    
+    do variableIndex = 1, dimNemovar
+      select case(NEMOvartype(variableIndex))
+        case('SC')
+          call utl_checkNetCDFstatus(nf90_def_var(ncid, NEMOvarnames(variableIndex), &
+	                                          nf90_double,           NEMOvarid(variableIndex)))
+        case('2D')
+          call utl_checkNetCDFstatus(nf90_def_var(ncid, NEMOvarnames(variableIndex), &
+	                                          nf90_float ,  dimID2D, NEMOvarid(variableIndex)))
+        case('3D')
+          call utl_checkNetCDFstatus(nf90_def_var(ncid, NEMOvarnames(variableIndex), &
+	                                          nf90_double,  dimID3D, NEMOvarid(variableIndex)))
+        case('4D')
+          call utl_checkNetCDFstatus(nf90_def_var(ncid, NEMOvarnames(variableIndex), &
+	                                          nf90_double,  dimID4D, NEMOvarid(variableIndex)))
+        case('TI')
+          call utl_checkNetCDFstatus(nf90_def_var(ncid, NEMOvarnames(variableIndex), &
+	                                          nf90_double,   dimIDTI, NEMOvarid(variableIndex)))
+	case('DE')
+          call utl_checkNetCDFstatus(nf90_def_var(ncid, NEMOvarnames(variableIndex), &
+	                                          nf90_float , dimIDDE, NEMOvarid(variableIndex)))
+      end select
+    end do
     
     ! end definitions
     call utl_checkNetCDFstatus(nf90_enddef(ncid))